cast to remove windows compiler warning
git-svn-id: https://svn.resiprocate.org/rep/resiprocate@4752 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ main/sip/resiprocate/os/AbstractFifo.cxx
@@ -45,7 +45,7 @@ AbstractFifo::getNext(int ms)
    while (mFifo.empty())
    {
       // bail if total wait time exceeds limit
-      bool signaled = mCondition.wait(mMutex, end - Timer::getTimeMs());
+      bool signaled = mCondition.wait(mMutex, (unsigned int)(end - Timer::getTimeMs()));
       if (!signaled)
       {
         return 0;
